RL: change float to double if necessary since float can be wrong with millions of adding.

+++ global.cpp
@@ -20,12 +20,12 @@ int Paras2probe::AllocateMemory(int n_file)
     Otime = new float[n_file];
     N_par = new long[n_file];
     Max_Rhop = new float[n_file];
-    RpAV = new float[n_file];
-    RpSQ = new float[n_file];
-    RpQU = new float[n_file];
-    Hp = new float[n_file];
+    RpAV = new double[n_file];
+    RpSQ = new double[n_file];
+    RpQU = new double[n_file];
+    Hp = new double[n_file];
     Hp_in1sigma = new float[n_file];
-    dSigma = new float[n_file];
+    dSigma = new double[n_file];
     for (int i = 0; i != n_file; i++) {
         // initialize if you don't assign all of them values but use them for calculation eventually
         Otime[i] = 0;
@@ -39,24 +39,24 @@ int Paras2probe::AllocateMemory(int n_file)
         dSigma[i] = 0;
     }
     
-    MeanSigma = new float*[n_file];
-    VpecG = new float*[n_file];
-    VertRho = new float*[n_file];
+    MeanSigma = new double*[n_file];
+    VpecG = new double*[n_file];
+    VertRho = new double*[n_file];
     CorrL = new float*[n_file];
 #ifdef CorrValue
     CorrV = new float*[n_file];
 #endif
-    GasHst = new float*[n_file];
-    GasHst2 = new float*[n_file];
-    ParHst = new float*[n_file];
-    ParLis = new float*[n_file];
+    GasHst = new double*[n_file];
+    GasHst2 = new double*[n_file];
+    ParHst = new double*[n_file];
+    ParLis = new double*[n_file];
     GPME = new double*[n_file];
     GPMEPar = new double*[n_file];
     for (int i = 0; i != n_file; i++) {
-        GasHst[i] = new float[16];
-        GasHst2[i] = new float[16];
-        ParHst[i] = new float[16];
-        ParLis[i] = new float[20];
+        GasHst[i] = new double[16];
+        GasHst2[i] = new double[16];
+        ParHst[i] = new double[16];
+        ParLis[i] = new double[20];
         for (int j = 0; j != 16; j++) {
             GasHst[i][j] = 0;
             GasHst2[i][j] = 0;
@@ -77,15 +77,15 @@ int Paras2probe::AllocateMemory(int n_file)
 int Paras2probe::AllocateSubMemory(int n_file, int *dimensions)
 {
     for (int i = 0; i != n_file; i++) {
-        MeanSigma[i] = new float[4*dimensions[0]];
+        MeanSigma[i] = new double[4*dimensions[0]];
         for (int j = 0; j != 4*dimensions[0]; j++) {
             MeanSigma[i][j] = 0;
         }
-        VpecG[i] = new float[3*dimensions[2]];
+        VpecG[i] = new double[3*dimensions[2]];
         for (int j = 0; j != 3*dimensions[2]; j++) {
             VpecG[i][j] = 0;
         }
-        VertRho[i] = new float[2*dimensions[2]];
+        VertRho[i] = new double[2*dimensions[2]];
         for (int j = 0; j != 2*dimensions[2]; j++) {
             VertRho[i][j] = 0;
         }
